+++ src/app/pages/admin/add-user-to-workgroup/page.js
@@ -3,7 +3,9 @@ import Layout from "@/components/Layout";
 import TableComponent from "@/components/TableComponent";
 import { getSession } from "@/lib/utils/utils";
 import { useEffect, useState } from "react";
-import {config} from "../../../../config/config.js";
+import { config } from "../../../../config/config.js";
+import Swal from "sweetalert2";
+import Image from "next/image";
 
 const workgroupHeader = ["id", "EMP_number", "Email", "Name", "Role", "Action"];
 const userHeader = ["id", "EMP_number", "Email", "Name", "Role", "Action"];
@@ -42,7 +44,7 @@ const Page = () => {
   const fetchUser = async (user_id) => {
     try {
       const response = await fetch(
-        `${config.host}/api/user/get-user/${user_id}`
+        `/api/user/get-user/${user_id}`, { next: { revalidate: 10 } }
       );
       if (!response.ok) {
         throw new Error("Failed to fetch user data");
@@ -59,7 +61,7 @@ const Page = () => {
   const fetchUsersWorkgroup = async (workgroup_id) => {
     try {
       const response = await fetch(
-        `${config.host}/api/workgroup/get-users-from-workgroup/${workgroup_id}`
+        `/api/workgroup/get-users-from-workgroup/${workgroup_id}`, { next: { revalidate: 10 } }
       );
       if (!response.ok) {
         throw new Error("Failed to fetch users");
@@ -75,13 +77,14 @@ const Page = () => {
   const fetchUsers = async () => {
     try {
       const response = await fetch(
-        `${config.host}/api/user/get-users`
+        `/api/user/get-users`, { next: { revalidate: 10 } }
       );
       if (!response.ok) {
         throw new Error("Failed to fetch users");
       }
       const usersData = await response.json();
-      const filteredUsers = usersData.users.filter(user => user.role !== "SA" && user.role !== "Admin Group");
+      //don't show if user already has some role
+      const filteredUsers = usersData.users.filter(user => user.role === "No role");
       setUsers(filteredUsers);
 
     } catch (error) {
@@ -91,11 +94,12 @@ const Page = () => {
 
   const fetchRoles = async () => {
     try {
-      const response = await fetch(`${config.host}/api/role/get-roles`, {
+      const response = await fetch(`/api/role/get-roles`, {
         method: 'GET',
         headers: {
           'Content-Type': 'application/json'
         },
+        next: { revalidate: 10 }
       });
 
       if (!response.ok) {
@@ -136,7 +140,7 @@ const Page = () => {
   const fetchUserEnabledFunction = async (user_id) => {
     try {
       const response = await fetch(
-        `${config.host}/api/action/get-user-action/${user_id}`
+        `/api/action/get-user-action/${user_id}`, { next: { revalidate: 10 } }
       );
       if (!response.ok) {
         throw new Error("Failed to fetch user data");
@@ -152,7 +156,7 @@ const Page = () => {
   const handleAdd = async (user_id, role_id) => {
     const workgroup_id = user.workgroup_id;
     // Update user role
-    const res = await fetch(`${config.host}/api/user/update-user/${user_id}`, {
+    const res = await fetch(`/api/user/update-user/${user_id}`, {
       method: "PUT",
       headers: {
         "Content-Type": "application/json",
@@ -160,11 +164,12 @@ const Page = () => {
       body: JSON.stringify({
         ROLE: role_id,
       }),
+      next: { revalidate: 10 }
     });
 
 
     // Add user to workgroup
-    await fetch(`${config.host}/api/workgroup/add-user-to-workgroup-admin`, {
+    await fetch(`/api/workgroup/add-user-to-workgroup-admin`, {
       method: "POST",
       headers: {
         "Content-Type": "application/json",
@@ -173,6 +178,7 @@ const Page = () => {
         user_id: user_id,
         workgroup_id: workgroup_id,
       }),
+      next: { revalidate: 10 }
     });
 
     setRefresh(!refresh);
@@ -181,20 +187,61 @@ const Page = () => {
   const handleRemove = async (user_id) => {
     const workgroup_id = user.workgroup_id;
 
-    // Remove user from workgroup
-    await fetch(`${config.host}/api/workgroup/remove-user-from-workgroup`, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
+    const swalWithBootstrapButtons = Swal.mixin({
+      customClass: {
+        confirmButton: "btn btn-success",
+        cancelButton: "btn btn-danger"
       },
-      body: JSON.stringify({
-        user_id: user_id,
-        workgroup_id: workgroup_id,
-      }),
+      buttonsStyling: true
     });
 
-    setRefresh(!refresh);
-  }
+    swalWithBootstrapButtons.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonText: "Yes, remove!",
+      cancelButtonText: "No, cancel!",
+      reverseButtons: true
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        // Remove user from workgroup
+        try {
+          await fetch(`/api/workgroup/remove-user-from-workgroup`, {
+            method: "DELETE",
+            headers: {
+              "Content-Type": "application/json",
+            },
+            body: JSON.stringify({
+              user_id: user_id,
+              workgroup_id: workgroup_id,
+            }),
+            next: { revalidate: 10 }
+          });
+          swalWithBootstrapButtons.fire({
+            title: "Removed!",
+            text: "The user has been removed from the workgroup.",
+            icon: "success"
+          });
+          setRefresh(!refresh);
+        } catch (error) {
+          console.error("Error removing user from workgroup:", error);
+        }
+      } else if (
+        /* Read more about handling dismissals below */
+        result.dismiss === Swal.DismissReason.cancel
+      ) {
+        swalWithBootstrapButtons.fire({
+          title: "Cancelled",
+          text: "The removal action was cancelled.",
+          icon: "error"
+        });
+      }
+    });
+  };
+
+
+
 
 
   const dataUsers = users
@@ -209,7 +256,7 @@ const Page = () => {
         Name: user.name,
         Role: <RoleSelect user_id={user._id} />,
         action: [
-          <span className="pl-4" key={index}>
+          <span className="pl-4 flex justify-center items-center" key={index}>
             <button
               onClick={() => {
                 const role_id = selectedRoles[user._id];
@@ -255,16 +302,26 @@ const Page = () => {
 
 
   return (
-    <Layout className="w-full h-screen flex flex-col gap-4 items-center justify-start font-sans">
-      <div className="w-full h-full bg-white container px-8  rounded-lg flex flex-col gap-8">
-        <h1 className="text-2xl font-bold text-primary flex  items-center">{">"} {user.workgroup} </h1>
-        <h1 className="text-1xl font-semibold">Add User to Workgroup</h1>
-        <div className="mt-4">
-          <TableComponent headers={workgroupHeader} datas={dataUsersWorkgroup} searchColumn={"Name"} TableName={"Members"}/>
+    <Layout className="container flex flex-col left-0 right-0 mx-auto justify-start font-sans mt-2">
+      <div className="flex flex-col items-start gap-4 mb-4 p-4 bg-white ">
+        <div className="flex items-center gap-4">
+          <Image src="/assets/card-logo/workgroup.png" alt="wd logo" width={50} height={50} />
+          <h1 className="text-3xl font-bold text-slate-900">ChecklistPM-Workgroup</h1>
+        </div>
+        <h1 className="text-sm font-bold text-secondary flex  items-center">Manage workgroup information, and assign users to each workgroup.</h1>
+      </div>
+
+
+      <div className="w-full h-full bg-white container  rounded-lg flex flex-col gap-4">
+        <h1 className="text-2xl font-bold text-primary flex  items-center">Current Workgroup: {user.workgroup} </h1>
+        <h1 className="text-sm font-bold text-secondary flex  items-center">Please select a role for the user before adding them to the workgroup.</h1>
+        <hr className="w-full" />
+        <div className="">
+          <TableComponent headers={workgroupHeader} datas={dataUsersWorkgroup} searchColumn={"Name"} TableName={"Members"} />
         </div>
         <hr className="w-full" />
-        <div className="mt-4">
-          <TableComponent headers={userHeader} datas={dataUsers} searchColumn={"Name"} TableName={"All users"}/>
+        <div className="">
+          <TableComponent headers={userHeader} datas={dataUsers} searchColumn={"Name"} TableName={"All users"} />
         </div>
       </div>
     </Layout>
